added all my stuff

+++ DigitalSecurityHub/products/views.py
@@ -4,9 +4,8 @@ from accounts.views import output_seller
 from django.http import HttpResponse, HttpResponseRedirect, JsonResponse
 from cart.models import Cart
 from accounts.models import Customer
+from django.db.models import Q
 import json
-# from django.views import ListView
-
 
 from .models import Product
 
@@ -19,38 +18,42 @@ def ProductList(request):
     """
     
     products = Product.objects.all()
+    
+    # Returns a list of all current products
     if request.method == 'GET':
         product_list = []
 
         for item in products:
             product_list.append(output_product(item))
         
-        return JsonResponse(product_list, safe=False)
+        # render a list of all the product titles
+        return render(request, 'products/productList.html', {'object_list': product_list}, status=200)
     
-    if request.method is 'POST':
+    # Allows a vendor to add a new product
+    if request.method is 'POST' and request.user.is_authenticated and Customer.objects.get(customer_id=request.user).type:
         try:
             json_post = json.loads(request.body)
             Product.objects.create(
                 title = json_post['title'],
                 description = json_post['description'],
                 price = json_post['price'],
-                seller = json_post['seller'],
                 stock = json_post['stock'],
                 active = json_post['active']
             )
             return HttpResponse('Product added')
-
         except:
-            return HttpResponse("Failed to process json.", status=500)
+            return HttpResponse("Failed to add new product.", status=500)
 
-    if request.method is 'DELETE':
+    # Allows the current user to delete all products owned by them
+    if request.method is 'DELETE' and request.user.is_authenticated and Customer.objects.get(customer_id=request.user).type:
         try:
-            json_post = json.loads(request.body)
-            Product.objects.get(id=id).delete()
+            seller = Customer.objects.get(seller_id=request.user)
+            product_list = Product.objects.get(seller_id=seller)  # <-- can you just put .delete() here to delet all products?
+            for item in product_list:
+                Product.objects.get(item).delete()  # <-- not actually sure if this is properly retrieving all objects
             return HttpResponse('Product deleted')
-
         except:
-            return HttpResponse("Failed to delete product.", status=500)
+            return HttpResponse("Failed to delete products.", status=500)
 
 
 def SpecificProduct(request, quantity, product_id):
@@ -61,36 +64,30 @@ def SpecificProduct(request, quantity, product_id):
     DELETE: deletes product if seller
     """
     
+    # Returns product information for the specified product
     if request.method is 'GET':
         try:
             json_post = json.loads(request.body)
-            product = Product.objects.get(id=product_id)
-            return JsonResponse(product, safe=False)
+            product = output_product(Product.objects.get(id=product_id))  # <-- using output product to format it correctly
+            return render(request, 'products/product.html', {'product': product})
         except:
             return HttpResponse('Could not find product')
         
-    
-    if request.method is 'POST':
+    # Should add the product to the cart of the authenticated user
+    if request.method is 'POST' and request.user.is_authenticated:
         try:
             json_post = json.loads(request.body)
-            product = Product.objects.get(
-                title = json_post['title'],
-                description = json_post['description'],
-                price = json_post['price'],
-                seller = json_post['seller'],
-                stock = json_post['stock'],
-                active = json_post['active']
-            )
+            product = Product.objects.get(id=product_id)  # <-- do not think we need to say which seller it is
             Cart.object.create(product=product, quantity=quantity, customer_id=request.user)
-            return HttpResponse('Product added')
+            return HttpResponse('Product added')  # <-- product should be added to the current logged in user's cart
         except:
             return HttpResponse("Could not add product to cart.", status=500)
 
-
-    if request.method is 'PATCH':
+    # Allows seller to edit product information (if it is their product)
+    if request.method is 'PATCH' and request.user.is_authenticated and Customer.objects.get(customer_id=request.user).type:
         try:
             json_post = json.loads(request.body)
-            product = Product.objects.get(id=product_id)
+            product = Product.objects.get(Q(id=product_id) & Q(seller_id=request.user))  # <-- should get the specified product of the current logged in seller
             product.title = json_post['title']
             product.description = json_post['description']
             product.price = json_post['price']
@@ -102,21 +99,22 @@ def SpecificProduct(request, quantity, product_id):
         except:
             return HttpResponse('Product could not be updated')
 
-
-    if request.method is 'DELETE':
+    # Will delete the specified product if owned by the current seller
+    if request.method is 'DELETE' and request.user.is_authenticated:
         if Customer.objects.get(customer_id=request.user).type:
             Product.objects.get(id=product_id).delete()
             return HttpResponse('Product deleted from store')
         else:
             return HttpResponse('You do not have the required permissions', type=401)
-    template_name = "products/list.html"
+    
+    return HttpResponse('Method not allowed', status=405)
 
-def product_list_view(request):
-    queryset = Product.objects.all()
-    context = {
-        'object_list': queryset
-    }
-    return render(request, "products/list.html", context)
+# def product_list_view(request):
+#     queryset = Product.objects.all()
+#     context = {
+#         'object_list': queryset
+#     }
+#     return render(request, "products/list.html", context)
 
 def output_product(product):
     """
